extract util functions into proper glsl utils modules
